pyproject.toml
[tool.poetry]
name = "worktime"
version = "0.1.0"
description = ""
authors = ["Pedro Silva <9375141+pedrox-hs@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
pandas = "^2.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
